@@ -1,11 +1,11 @@
 ext {
     //android开发版本配置
     android = [
-            compileSdkVersion: 28,
-            buildToolsVersion: "28.0.0",
+            compileSdkVersion: 29,
+            buildToolsVersion: "29.0.0",
             applicationId    : "com.experiment.npe",
-            minSdkVersion    : 21,
-            targetSdkVersion : 28,
+            minSdkVersion    : 26,
+            targetSdkVersion : 29,
             versionCode      : 1,
             versionName      : "1.0",
     ]
@@ -17,12 +17,12 @@ ext {
     //support配置
     support = [
             "constraint-layout"       : "1.0.0-beta2",
-            'support-v4'              : "com.android.support:support-v4:${versions["support-version"]}",
-            'appcompat-v7'            : "com.android.support:appcompat-v7:${versions["support-version"]}",
-            'recyclerview-v7'         : "com.android.support:recyclerview-v7:${versions["support-version"]}",
+            'support-v4'              : 'androidx.legacy:legacy-support-v4:1.0.0',
+            'appcompat-v7'            : 'androidx.appcompat:appcompat:1.0.0',
+            'recyclerview-v7'         : 'androidx.recyclerview:recyclerview:1.0.0',
             'support-v13'             : "com.android.support:support-v13:${versions["support-version"]}",
             'support-fragment'        : "com.android.support:support-fragment:${versions["support-version"]}",
-            'design'                  : "com.android.support:design:${versions["support-version"]}",
+            'design'                  : 'com.google.android.material:material:1.0.0',
             'animated-vector-drawable': "com.android.support:animated-vector-drawable:${versions["support-version"]}",
             'junit'                   : "junit:junit:${versions["junit-version"]}",
     ]
@@ -55,8 +55,8 @@ ext {
             "bindingcollectionadapter"             : "me.tatarka.bindingcollectionadapter2:bindingcollectionadapter:2.2.0",
             "bindingcollectionadapter-recyclerview": "me.tatarka.bindingcollectionadapter2:bindingcollectionadapter-recyclerview:2.2.0",
             //Google AAC
-            "lifecycle-extensions"                 : "android.arch.lifecycle:extensions:1.1.1",
-            "lifecycle-compiler"                 : "android.arch.lifecycle:compiler:1.1.1",
+            "lifecycle-extensions"                 : 'androidx.lifecycle:lifecycle-extensions:2.0.0',
+            "lifecycle-compiler"                 : 'androidx.lifecycle:lifecycle-compiler:2.0.0',
             //MVVMHabit
             "MVVMHabit"                            : "com.github.goldze:MVVMHabit:3.1.6",
             "PictureSelector"                       : "com.github.LuckSiege.PictureSelector:picture_library:v2.5.9",
