@@ -1,13 +1,11 @@
 package com.experiment.npe.app;
 
 import android.annotation.SuppressLint;
-import android.app.Activity;
 import android.app.Application;
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.support.annotation.NonNull;
-import android.support.annotation.VisibleForTesting;
-import android.support.v4.app.Fragment;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.annotation.NonNull;
+import androidx.annotation.VisibleForTesting;
 
 import com.experiment.npe.data.NpeRepository;
 import com.experiment.npe.ui.login.LoginViewModel;
@@ -15,7 +13,6 @@ import com.experiment.npe.ui.main.viewmodel.TabBar1ViewModel;
 import com.experiment.npe.ui.main.viewmodel.TabBar2ViewModel;
 import com.experiment.npe.ui.regist.RegistViewModel;
 import com.experiment.npe.ui.search.SearchViewModel;
-import com.experiment.npe.ui.setting.SettingViewModel;
 
 
 /**
