@@ -1,9 +1,9 @@
 package com.experiment.npe.ui.regist;
 
 import android.app.Application;
-import android.databinding.ObservableField;
-import android.databinding.ObservableInt;
-import android.support.annotation.NonNull;
+import androidx.databinding.ObservableField;
+import androidx.databinding.ObservableInt;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 import android.view.View;
 
