@@ -1,17 +1,13 @@
 package com.experiment.npe.ui.search;
 
-import android.arch.lifecycle.ViewModelProviders;
-import android.support.v7.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 
 import com.experiment.npe.BR;
 import com.experiment.npe.R;
 import com.experiment.npe.app.AppViewModelFactory;
-import com.experiment.npe.databinding.ActivityLoginBinding;
 import com.experiment.npe.databinding.ActivitySearchBinding;
-import com.experiment.npe.ui.login.LoginViewModel;
-import com.experiment.npe.ui.regist.RegistViewModel;
 
 import me.goldze.mvvmhabit.base.BaseActivity;
 
