@@ -1,15 +1,11 @@
 package com.experiment.npe.ui.setting;
 
-import android.arch.lifecycle.ViewModelProviders;
-import android.support.v7.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.experiment.npe.R;
-import com.experiment.npe.app.AppViewModelFactory;
-import com.experiment.npe.databinding.ActivityLoginBinding;
-import com.experiment.npe.databinding.ActivitySearchBinding;
 import com.experiment.npe.databinding.ActivitySettingBinding;
-import com.experiment.npe.ui.login.LoginViewModel;
 
 import me.goldze.mvvmhabit.base.BaseActivity;
 import me.tatarka.bindingcollectionadapter2.BR;
