@@ -1,7 +1,7 @@
 package me.goldze.mvvmhabit.base;
 
 import android.app.Activity;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import java.util.Stack;
 
