@@ -3,7 +3,7 @@ package me.goldze.mvvmhabit.base;
 import android.app.Activity;
 import android.app.Application;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import me.goldze.mvvmhabit.utils.Utils;
 
