@@ -1,9 +1,8 @@
 package me.goldze.mvvmhabit.binding.viewadapter.radiogroup;
 
-import android.databinding.BindingAdapter;
-import android.support.annotation.IdRes;
-import android.text.TextUtils;
-import android.webkit.WebView;
+import androidx.databinding.BindingAdapter;
+import androidx.annotation.IdRes;
+
 import android.widget.RadioButton;
 import android.widget.RadioGroup;
 
