@@ -1,7 +1,7 @@
 package me.goldze.mvvmhabit.binding.viewadapter.swiperefresh;
 
-import android.databinding.BindingAdapter;
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.databinding.BindingAdapter;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import me.goldze.mvvmhabit.binding.command.BindingCommand;
 
