@@ -2,7 +2,7 @@ package me.goldze.mvvmhabit.utils;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 /**
  * Created by goldze on 2017/5/14.
