@@ -1,6 +1,6 @@
 package me.goldze.mvvmhabit.binding.viewadapter.checkbox;
 
-import android.databinding.BindingAdapter;
+import androidx.databinding.BindingAdapter;
 import android.widget.CheckBox;
 import android.widget.CompoundButton;
 
