@@ -1,7 +1,7 @@
 package me.goldze.mvvmhabit.binding.viewadapter.edittext;
 
 import android.content.Context;
-import android.databinding.BindingAdapter;
+import androidx.databinding.BindingAdapter;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.inputmethod.InputMethodManager;
