@@ -1,16 +1,13 @@
 package me.goldze.mvvmhabit.binding.viewadapter.image;
 
 
-import android.databinding.BindingAdapter;
+import androidx.databinding.BindingAdapter;
 import android.text.TextUtils;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.resource.bitmap.CircleCrop;
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners;
 import com.bumptech.glide.request.RequestOptions;
-
-import me.goldze.mvvmhabit.R;
 
 /**
  * Created by goldze on 2017/6/18.
