@@ -1,15 +1,13 @@
 package me.goldze.mvvmhabit.binding.viewadapter.viewgroup;
 
-import android.databinding.BindingAdapter;
-import android.databinding.DataBindingUtil;
-import android.databinding.ObservableList;
-import android.databinding.ViewDataBinding;
+import androidx.databinding.BindingAdapter;
+import androidx.databinding.DataBindingUtil;
+import androidx.databinding.ObservableList;
+import androidx.databinding.ViewDataBinding;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 
 import me.tatarka.bindingcollectionadapter2.ItemBinding;
-
-import static me.goldze.mvvmhabit.R.attr.itemView;
 
 /**
  * Created by goldze on 2017/6/18.
